requirements.txt
requests==2.23.0
python-dotenv==0.13.0
pyTelegramBotAPI==3.7.1
schedule==0.6.0
